move the SD Card module to src as repository
