+++ src/service/SyncService.h
@@ -1,9 +1,9 @@
 #ifndef SYNC_SERVICE_H
 #define SYNC_SERVICE_H
 
-#include "SDCardModule.h"
-#include <esp_log.h>
+#include "repository/SDCardModule/SDCardModule.h"
 #include "communication/ble/BLEModule.h"
+#include <esp_log.h>
 
 class SyncService {
     public:
